check-access-token-route

+++ Front-End/Scripts/profile.js
@@ -4,7 +4,7 @@ const BaseUrl_profile = `https://angry-cummerbund-newt.cyclic.app`
 
 const qrcodeuserdetails_qrpage11 = localStorage.getItem('qrcodeuserdetails') || null;
 // console.log(qrcodeuserdetails_qrpage);
-if(!qrcodeuserdetails_qrpage11){
+if (!qrcodeuserdetails_qrpage11) {
     alert('Kindly Login First');
     location.href = '../View/login.html'
 }
@@ -30,7 +30,7 @@ const userGender = document.getElementById("login-user-gender")
 
 function populate() {
     const details = JSON.parse(localStorage.getItem("qrcodeuserdetails")) || null
-    console.log(details)
+    // console.log(details)
 
     if (details) {
         userNameHeading.innerText = details.Name
@@ -46,50 +46,71 @@ function populate() {
 
 
 function forgotPasswordBtnProfile() {
-    console.log("update")
+    // console.log("update")
     location.href = "./forgetpass.html"
 }
 
 
-function updateProfile(){
-    console.log('update click');
+function updateProfile() {
 
     const payload = {
-        Name : userName.value,
-        Address : userAddress.value,
-        Gender : userGender.value
+        Name: userName.value,
+        Address: userAddress.value,
+        Gender: userGender.value
     }
 
     const detailsforuserid = JSON.parse(localStorage.getItem("qrcodeuserdetails")) || null
-    if(!detailsforuserid){
-        alert('login first')
+    if (!detailsforuserid) {
+        alert('Please Login Again.(Session Time Out)')
         return
     }
 
-    if(!detailsforuserid._id){
-        alert('login first')
+    if (!detailsforuserid._id) {
+        alert('Please Login Again.(Session Time Out)')
         return
     }
 
-    
-    fetch(`${BaseUrl_profile}/profile/updateProfile/${detailsforuserid._id}`, {
-        method : "PATCH",
-        headers : {
-            "Content-type" : "application/json",
-        },
-        body : JSON.stringify( payload )
-    })
-    .then(res => res.json() )
-    .then(data => {
-        if(data?.user){
+    document.getElementById('user-profile-updateBtn').innerHTML = `<i class="fa fa-refresh fa-spin"></i> Update Profile`;
+    document.getElementById('user-profile-updateBtn').disabled = true;
 
-            localStorage.setItem('qrcodeuserdetails', JSON.stringify(data.user))
-        }
-        window.location.reload()
-    }).catch(err => {
-        console.error(err);
-        alert('something went wrong. try after some time')
-    })
+
+
+    if (confirm('Do You Want To Update Your Profile?')) {
+        fetch(`${BaseUrl_profile}/profile/updateProfile/${detailsforuserid._id}`, {
+            method: "PATCH",
+            headers: {
+                "Content-type": "application/json",
+            },
+            body: JSON.stringify(payload)
+        })
+            .then(res => res.json())
+            .then(data => {
+                document.getElementById('user-profile-updateBtn').innerHTML = `Update Profile`;
+                document.getElementById('user-profile-updateBtn').disabled = false;
+
+                console.log(data);
+
+                if (data?.user) {
+
+                    localStorage.setItem('qrcodeuserdetails', JSON.stringify(data.user))
+
+                }
+                // window.location.reload()
+            }).catch(err => {
+                console.error(err);
+
+                document.getElementById('user-profile-updateBtn').innerHTML = `Update Profile`;
+                document.getElementById('user-profile-updateBtn').disabled = false;
+
+                alert('Something Went Wrong.(Please Try After Some Time)')
+            })
+    } else {
+        document.getElementById('user-profile-updateBtn').innerHTML = `Update Profile`;
+        document.getElementById('user-profile-updateBtn').disabled = false;
+    }
+
+
+
 
 
 
@@ -98,37 +119,37 @@ function updateProfile(){
 
 
 
-function deleteProfile(){
-    
+function deleteProfile() {
+
     const detailsforuserid = JSON.parse(localStorage.getItem("qrcodeuserdetails")) || null
-    if(!detailsforuserid){
+    if (!detailsforuserid) {
         alert('login first')
         return
     }
 
-    if(!detailsforuserid._id){
+    if (!detailsforuserid._id) {
         alert('login first')
         return
     }
 
-    
+
     fetch(`${BaseUrl_profile}/profile/deleteProfile/${detailsforuserid._id}`, {
-        method : "DELETE"
+        method: "DELETE"
     })
-    .then(res => res.json() )
-    .then(data => {
-        console.log(data);
-        if(data.success){
-            console.log('Your Accout Successfully Deleted');
-            localStorage.removeItem('qrcodeuserdetails');
-            location.href = '../index.html'
-        }else{
-            alert(`You can't Able to Delete Account Yet. Please Contact To Manager`)
-        }
-        
-    }).catch(err => {
-        console.error(err);
-        alert('something went wrong. try after some time')
-    })
+        .then(res => res.json())
+        .then(data => {
+            console.log(data);
+            if (data.success) {
+                console.log('Your Accout Successfully Deleted');
+                localStorage.removeItem('qrcodeuserdetails');
+                location.href = '../index.html'
+            } else {
+                alert(`You can't Able to Delete Account Yet. Please Contact To Manager`)
+            }
+
+        }).catch(err => {
+            console.error(err);
+            alert('something went wrong. try after some time')
+        })
 
 }
